requirements.txt
python-jose
urllib
